+++ rdt/__init__.py
@@ -4,7 +4,7 @@
 
 __author__ = 'DataCebo, Inc.'
 __email__ = 'info@sdv.dev'
-__version__ = '1.12.2'
+__version__ = '1.12.3.dev1'
 
 
 import sys