+++ tests/unit/transformers/pii/test_utils.py
@@ -1,4 +1,4 @@
-from rdt.transformers.pii.utils import get_provider_name, is_faker_function
+from rdt.transformers.pii.utils import get_provider_name
 
 
 def test_get_provider_name():
@@ -14,21 +14,3 @@ def test_get_provider_name():
     # Assert
     assert email_provider == 'internet'
     assert lexify_provider == 'BaseProvider'
-
-
-def test_is_faker_function():
-    """Test that the method returns True if the ``function_name`` is a valid faker function."""
-    # Run
-    result = is_faker_function('address')
-
-    # Assert
-    assert result is True
-
-
-def test_is_faker_function_error():
-    """Test that the method returns False if ``function_name`` is not a valid faker function."""
-    # Run
-    result = is_faker_function('blah')
-
-    # Assert
-    assert result is False